Extracted extension for Float4x4, UIGestureRecognizer & CGPoint
