+++ Sources/UI/ARView/ARView.swift
@@ -134,50 +134,3 @@ public extension ARView {
       self.viewDelegate?.arView(self, updatedPositionOf: virtualObject)
   }
 }
-
-// MARK: - float4x4 extensions
-private extension float4x4 {
-  /**
-   Treats matrix as a (right-hand column-major convention) transform matrix
-   and factors out the translation component of the transform.
-  */
-  var translation: SIMD3<Float> {
-    get {
-      let translation = columns.3
-      return [translation.x, translation.y, translation.z]
-    }
-    set(newValue) {
-      columns.3 = [newValue.x, newValue.y, newValue.z, columns.3.w]
-    }
-  }
-  
-  /**
-   Factors out the orientation component of the transform.
-  */
-  var orientation: simd_quatf {
-    return simd_quaternion(self)
-  }
-  
-  /**
-   Creates a transform matrix with a uniform scale factor in all directions.
-   */
-  init(uniformScale scale: Float) {
-    self = matrix_identity_float4x4
-    columns.0.x = scale
-    columns.1.y = scale
-    columns.2.z = scale
-  }
-}
-
-// MARK: - CGPoint extensions
-private extension CGPoint {
-  /// Extracts the screen space point from a vector returned by SCNView.projectPoint(_:).
-  init(_ vector: SCNVector3) {
-    self.init(x: CGFloat(vector.x), y: CGFloat(vector.y))
-  }
-
-  /// Returns the length of a point when considered as a vector. (Used with gesture recognizers.)
-  var length: CGFloat {
-    return sqrt(x * x + y * y)
-  }
-}
