+++ Sources/UI/ARView/VirtualObjectInteraction.swift
@@ -163,26 +163,3 @@ extension VirtualObjectInteraction: UIGestureRecognizerDelegate {
     true
   }
 }
-
-// MARK: - UIGestureRecognizer extensions
-private extension UIGestureRecognizer {
-  func center(in view: UIView) -> CGPoint? {
-    guard numberOfTouches > 0 else { return nil }
-    
-    let first = CGRect(origin: location(ofTouch: 0, in: view), size: .zero)
-
-    let touchBounds = (1..<numberOfTouches).reduce(first) { touchBounds, index in
-      return touchBounds.union(CGRect(origin: location(ofTouch: index, in: view), size: .zero))
-    }
-
-    return CGPoint(x: touchBounds.midX, y: touchBounds.midY)
-  }
-}
-
-// MARK: - CGPoint extensions
-private extension CGPoint {
-  /// Extracts the screen space point from a vector returned by SCNView.projectPoint(_:).
-  init(_ vector: SCNVector3) {
-    self.init(x: CGFloat(vector.x), y: CGFloat(vector.y))
-  }
-}
